+++ test.js
@@ -20,12 +20,21 @@ assert.equal(strand.buffer.readUInt8((5 * 3) + 2), 3);
 var createStream = require("./index");
 var stream = createStream();
 stream.writePixels(0, strand.buffer);
+stream.writeMessage(5, 10, new Buffer([0x0, 0x1, 0x2]));
 stream.end();
 stream.pipe(concat(function(data) {
+  // .writePixels(0, strand.buffer);
   assert.equal(data.readUInt8(0), 0);
   assert.equal(data.readUInt8(1), 0);
   assert.equal(data.readUInt16BE(2), strand.buffer.length);
-  assert(bufferEqual(data.slice(4), strand.buffer));
+  assert(bufferEqual(data.slice(4, 4 + strand.buffer.length), strand.buffer));
+
+  // .writeMessage(5, 10, new Buffer([0x0, 0x1, 0x2]));
+  var offset = strand.buffer.length + 4;
+  assert.equal(data.readUInt8(offset + 0), 5);
+  assert.equal(data.readUInt8(offset + 1), 10);
+  assert.equal(data.readUInt16BE(offset + 2), 3);
+  assert(bufferEqual(data.slice(offset + 4), new Buffer([0x0, 0x1, 0x2])));
 }));
 
 // @TODO Test stream.writeColorCorrection