+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Next Release
+
+ * Add `stream.writeMessage`
+
 ## 1.0.0
 
  * Open Pixel Control protocol stream